requirements.txt
flask
cryptography
PyJWT
requests
